Add files via upload
Optimized the loading time of the program to request the URL of the database file and download it only if it doesn't exist in the folder.

Ensured that Pyspark reads the title of the movie inside quotation marks in case a title has a comma.

Turned inputs of "adult" column into boolean values to filter them out of outputs.

Added printing for first 20 movies with revenue over 1 billion and top 10 movies with the highest revenue. For the latter, added a filter to show only movies with vote count of over 1 to exclude incorrect data.